Python2.x support for maps
This commit introduces serialization and deserialization for the new Erlang map
type. The Erlang map type doesn't correspond very well to Python's native dict
type as Erlang's lists and maps are are hashable. In order to deal with this
behavioural mismatch, the List and ImproperList types have been made immutable
and a new immutable Map type has been introduced.